Extract marker detail modal flows
