+++ tests/test-marker-value-notes.js
@@ -85,14 +85,15 @@ const state = (await import('../js/state.js')).state;
   console.log('%c 4. saveManualEntry stores note ', 'font-weight:bold;color:#f59e0b');
 
   const viewsSrc = read('js/views.js');
+  const markerDetailSrc = read('js/marker-detail-modal.js');
   assert('saveManualEntry reads me-note from the form',
-    /const\s+noteInput\s*=\s*document\.getElementById\('me-note'\)/.test(viewsSrc));
+    /const\s+noteInput\s*=\s*document\.getElementById\('me-note'\)/.test(markerDetailSrc));
   assert('saveManualEntry stores noteText in markerValueNotes when non-empty',
-    /if \(noteText\) state\.importedData\.markerValueNotes\[noteKey\] = noteText/.test(viewsSrc));
+    /if \(noteText\) state\.importedData\.markerValueNotes\[noteKey\] = noteText/.test(markerDetailSrc));
   assert('saveManualEntry clears the entry when noteText is empty (idempotent edit-to-blank)',
-    /else delete state\.importedData\.markerValueNotes\[noteKey\]/.test(viewsSrc));
+    /else delete state\.importedData\.markerValueNotes\[noteKey\]/.test(markerDetailSrc));
   assert('manual-entry form HTML includes the me-note textarea',
-    viewsSrc.includes('id="me-note"') && /placeholder=".*fasted/i.test(viewsSrc));
+    markerDetailSrc.includes('id="me-note"') && /placeholder=".*fasted/i.test(markerDetailSrc));
 
   // ═══════════════════════════════════════
   // 5. editValueNote / deleteValueNote handlers
@@ -100,17 +101,17 @@ const state = (await import('../js/state.js')).state;
   console.log('%c 5. Value-note CRUD handlers ', 'font-weight:bold;color:#f59e0b');
 
   assert('editValueNote handler exported',
-    /export async function editValueNote\(id, date\)/.test(viewsSrc));
+    /export async function editValueNote\(id, date\)/.test(markerDetailSrc));
   assert('deleteValueNote handler exported',
-    /export async function deleteValueNote\(id, date\)/.test(viewsSrc));
+    /export async function deleteValueNote\(id, date\)/.test(markerDetailSrc));
   assert('editValueNote bound to window for inline onclicks',
     /editValueNote,\s*$/m.test(viewsSrc) || viewsSrc.includes('editValueNote,'));
   assert('deleteValueNote bound to window for inline onclicks',
     viewsSrc.includes('deleteValueNote,'));
   assert('editValueNote re-renders the detail modal on save',
-    /editValueNote[\s\S]{0,1500}showDetailModal\(id\)/.test(viewsSrc));
+    /editValueNote[\s\S]{0,1500}showDetailModal\(id\)/.test(markerDetailSrc));
   assert('deleteValueNote confirms before removing',
-    /deleteValueNote[\s\S]{0,400}showConfirmDialog\(/.test(viewsSrc));
+    /deleteValueNote[\s\S]{0,400}showConfirmDialog\(/.test(markerDetailSrc));
 
   // Direct state manipulation — verify the data model is what render code expects.
   state.importedData = state.importedData || {};
@@ -127,7 +128,7 @@ const state = (await import('../js/state.js')).state;
   console.log('%c 6. Orphan cleanup ', 'font-weight:bold;color:#f59e0b');
 
   assert('deleteMarkerValue drops the per-value note for the same (date, marker)',
-    /deleteMarkerValue[\s\S]{0,2000}delete state\.importedData\.markerValueNotes\[dotKey \+ ':' \+ date\]/.test(viewsSrc));
+    /deleteMarkerValue[\s\S]{0,2000}delete state\.importedData\.markerValueNotes\[dotKey \+ ':' \+ date\]/.test(markerDetailSrc));
 
   // Insulin dual-mapping parity: the value mirrors hormones.insulin ↔
   // diabetes.insulin_d, so the per-value note must mirror too. Bidirectional
@@ -136,27 +137,27 @@ const state = (await import('../js/state.js')).state;
 
   // 500-char cap defends against runaway paste (matches the wearable note cap).
   assert('saveManualEntry caps the note at 500 chars before storing',
-    /noteRaw\.length > 500 \? noteRaw\.slice\(0, 500\) : noteRaw/.test(viewsSrc));
+    /noteRaw\.length > 500 \? noteRaw\.slice\(0, 500\) : noteRaw/.test(markerDetailSrc));
   assert('editValueNote caps the note at 500 chars before storing',
-    /editValueNote[\s\S]{0,1200}result\.length > 500 \? result\.slice\(0, 500\) : result/.test(viewsSrc));
+    /editValueNote[\s\S]{0,1200}result\.length > 500 \? result\.slice\(0, 500\) : result/.test(markerDetailSrc));
 
   // editValueNote + deleteValueNote also route through _insulinMirrorNoteKey
   // — see the bidirectional helper asserts below.
   assert('deleteValueNote cleans the mirror note for insulin',
-    /deleteValueNote[\s\S]{0,800}_insulinMirrorNoteKey\(dotKey, date\)/.test(viewsSrc));
+    /deleteValueNote[\s\S]{0,800}_insulinMirrorNoteKey\(dotKey, date\)/.test(markerDetailSrc));
 
   // Greptile P1: insulin note mirror must be BIDIRECTIONAL — user may
   // edit/delete via the hormones panel OR the diabetes panel.
   assert('_insulinMirrorNoteKey helper defined and bidirectional',
-    /_insulinMirrorNoteKey\(dotKey, date\)/.test(viewsSrc) &&
-    /if \(dotKey === 'hormones\.insulin'\) return 'diabetes\.insulin_d:' \+ date/.test(viewsSrc) &&
-    /if \(dotKey === 'diabetes\.insulin_d'\) return 'hormones\.insulin:' \+ date/.test(viewsSrc));
+    /_insulinMirrorNoteKey\(dotKey, date\)/.test(markerDetailSrc) &&
+    /if \(dotKey === 'hormones\.insulin'\) return 'diabetes\.insulin_d:' \+ date/.test(markerDetailSrc) &&
+    /if \(dotKey === 'diabetes\.insulin_d'\) return 'hormones\.insulin:' \+ date/.test(markerDetailSrc));
   assert('saveManualEntry uses bidirectional mirror helper',
-    /saveManualEntry[\s\S]{0,2500}_insulinMirrorNoteKey\(dotKey, date\)/.test(viewsSrc));
+    /saveManualEntry[\s\S]{0,2500}_insulinMirrorNoteKey\(dotKey, date\)/.test(markerDetailSrc));
   assert('deleteMarkerValue uses bidirectional mirror helper',
-    /deleteMarkerValue[\s\S]{0,2500}_insulinMirrorNoteKey\(dotKey, date\)/.test(viewsSrc));
+    /deleteMarkerValue[\s\S]{0,2500}_insulinMirrorNoteKey\(dotKey, date\)/.test(markerDetailSrc));
   assert('editValueNote uses bidirectional mirror helper',
-    /editValueNote[\s\S]{0,1500}_insulinMirrorNoteKey\(dotKey, date\)/.test(viewsSrc));
+    /editValueNote[\s\S]{0,1500}_insulinMirrorNoteKey\(dotKey, date\)/.test(markerDetailSrc));
 
   // CodeQL js/xss-through-dom: empty-cell onclick must use JSON.stringify
   // so interpolated id/date survive the HTML-attr → JS-string round-trip.
@@ -169,13 +170,13 @@ const state = (await import('../js/state.js')).state;
   console.log('%c 7. Value-card render ', 'font-weight:bold;color:#f59e0b');
 
   assert('Value card reads note from markerValueNotes by mvKey',
-    /state\.importedData\.markerValueNotes\?\.\[mvKey\]/.test(viewsSrc));
+    /state\.importedData\.markerValueNotes\?\.\[mvKey\]/.test(markerDetailSrc));
   assert('Empty card shows "+ note" hint',
-    /mv-value-note add-note[\s\S]{0,200}\+ note/.test(viewsSrc));
+    /mv-value-note add-note[\s\S]{0,200}\+ note/.test(markerDetailSrc));
   assert('Populated card has × delete button',
-    /mv-value-note-delete[\s\S]{0,400}deleteValueNote\('/.test(viewsSrc));
+    /mv-value-note-delete[\s\S]{0,400}deleteValueNote\('/.test(markerDetailSrc));
   assert('Inline onclicks stopPropagation so cell-edit doesn\'t fire',
-    /editValueNote[\s\S]{0,200}event\.stopPropagation\(\)/.test(viewsSrc));
+    /editValueNote[\s\S]{0,200}event\.stopPropagation\(\)/.test(markerDetailSrc));
 
   // ═══════════════════════════════════════
   // 8. AI context emission — section:markerValueNotes
