+++ tests/verify-modules.js
@@ -559,6 +559,9 @@
     const hasViewsJs = sw.includes('/js/views.js');
     assert('Service worker caches js/views.js', hasViewsJs);
 
+    const hasMarkerDetailModalJs = sw.includes('/js/marker-detail-modal.js');
+    assert('Service worker caches js/marker-detail-modal.js', hasMarkerDetailModalJs);
+
     const hasSchemaJs = sw.includes('/js/schema.js');
     assert('Service worker caches js/schema.js', hasSchemaJs);
